Add TpmContext
TpmContext represents a TpmInstance which is created by the MS Tpm Lib.
Actually it is an object which contains all the global variables used
in MS Tpm Lib.

Fortunately these global variables are managed in below two header
files in MS Tpm Lib:
 - Global.h
 - PlatformData.h

So in Global.h we design a new data structure TPM_CONTEXT which
contains all the global variables. In TpmContext.c we implement
the save and pop functions to do the context-switch.

To make the implementation simple, we design a TPM_CONTEXT array to
manage the contexts. It can be refined to be a linked-list.

SwitchTpmContext is the function which does the context-switch. It
should be called before any external lib function call, for example,
ExecuteCommand.

Signed-off-by: Min Xu <min.m.xu@intel.com>